# Conflicts:
#	src/main/kotlin/no/nav/bidrag/behandling/service/UnderholdService.kt

+++ src/main/kotlin/no/nav/bidrag/behandling/service/UnderholdService.kt
@@ -61,12 +61,17 @@ class UnderholdService(
         gjelderBarn.validere()
 
         return gjelderBarn.personident?.let { personidentBarn ->
-            val rolle = behandling.søknadsbarn.find { it.ident == personidentBarn.verdi }
+            val rolleSøknadsbarn = behandling.søknadsbarn.find { it.ident == personidentBarn.verdi }
             personRepository.findFirstByIdent(personidentBarn.verdi)?.let {
                 lagreUnderholdskostnad(behandling, it)
             } ?: run {
-                val roller = rolle?.let { mutableSetOf(it) } ?: mutableSetOf()
-                val person = personRepository.save(Person(ident = personidentBarn.verdi, rolle = roller))
+                val person =
+                    personRepository.save(
+                        Person(
+                            ident = personidentBarn.verdi,
+                            rolle = rolleSøknadsbarn?.let { mutableSetOf(it) } ?: mutableSetOf(),
+                        ),
+                    )
                 person.rolle.forEach { it.person = person }
 
                 lagreUnderholdskostnad(behandling, person)
@@ -249,7 +254,11 @@ class UnderholdService(
         underholdskostnad: Underholdskostnad,
     ): UnderholdDto? {
         behandling.underholdskostnader.remove(underholdskostnad)
-        personRepository.deleteById(underholdskostnad.person.id!!)
+        underholdskostnad.person.underholdskostnad.remove(underholdskostnad)
+        if (underholdskostnad.person.underholdskostnad.isEmpty() && underholdskostnad.person.rolle.isEmpty()) {
+            personRepository.deleteById(underholdskostnad.person.id!!)
+            notatService.sletteNotat(behandling, Notattype.UNDERHOLDSKOSTNAD, behandling.bidragsmottaker!!)
+        }
         underholdskostnadRepository.deleteById(underholdskostnad.id!!)
         return null
     }
